+++ index.js
@@ -1,10 +1,12 @@
 const { getTree, getTreeForProject } = require("./src/lib/project-tree");
 const { readDefinitionFile } = require("./src/lib/reader");
 const { parentChainFromNode } = require("./src/lib/util/chain-util");
+const { treatUrl } = require("./src/lib/util/reader-util");
 
 module.exports = {
   getTree,
   getTreeForProject,
   readDefinitionFile,
-  parentChainFromNode
+  parentChainFromNode,
+  treatUrl
 };
