+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kie/build-chain-configuration-reader",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "description": "A library to read build-chain tool configuration files",
   "main": "index.js",
   "repository": {
