+++ test/lib/reader.test.js
@@ -15,9 +15,6 @@ test("readDefinitionFile with dependencies", async () => {
   );
   // Assert
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "test/resources/build-config-with-dependencies.yaml"
-  );
 });
 
 test("readDefinitionFile with external dependencies", async () => {
@@ -27,7 +24,6 @@ test("readDefinitionFile with external dependencies", async () => {
   );
   // Assert
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual("test/resources/build-config.yaml");
 });
 
 test("readDefinitionFile with external dependencies as URL", async () => {
@@ -43,9 +39,6 @@ test("readDefinitionFile with external dependencies as URL", async () => {
   );
   // Assert
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "test/resources/build-config-dependencies-url.yaml"
-  );
 });
 
 test("readDefinitionFileFromUrl relative embedded dependencies", async () => {
@@ -69,9 +62,6 @@ test("readDefinitionFileFromUrl relative embedded dependencies", async () => {
 
   expect(Array.isArray(result.dependencies)).toBe(true);
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "https://raw.githubusercontent.com/kiegroup/build-chain-configuration-reader/main/test/resources/build-config.yaml"
-  );
 });
 
 test("readDefinitionFileFromUrl relative path dependencies", async () => {
@@ -100,9 +90,6 @@ test("readDefinitionFileFromUrl relative path dependencies", async () => {
 
   expect(Array.isArray(result.dependencies)).toBe(true);
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "https://raw.githubusercontent.com/kiegroup/build-chain-configuration-reader/main/test/resources/build-config.yaml"
-  );
 });
 
 test("readDefinitionFileFromUrl url dependencies", async () => {
@@ -133,9 +120,6 @@ test("readDefinitionFileFromUrl url dependencies", async () => {
 
   expect(Array.isArray(result.dependencies)).toBe(true);
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "https://raw.githubusercontent.com/kiegroup/build-chain-configuration-reader/main/test/resources/build-config.yaml"
-  );
 });
 
 test("readDefinitionFileFromUrl with place holders relative path dependencies", async () => {
@@ -165,9 +149,6 @@ test("readDefinitionFileFromUrl with place holders relative path dependencies", 
 
   expect(Array.isArray(result.dependencies)).toBe(true);
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "https://raw.githubusercontent.com/groupx/projectx/branchx/test/resources/build-config.yaml"
-  );
 });
 
 test("readDefinitionFileFromUrl url dependencies placeholders", async () => {
@@ -204,7 +185,4 @@ test("readDefinitionFileFromUrl url dependencies placeholders", async () => {
 
   expect(Array.isArray(result.dependencies)).toBe(true);
   expect(result.dependencies.length).toEqual(25);
-  expect(result.sourceFile).toEqual(
-    "https://raw.githubusercontent.com/kiegroup/build-chain-configuration-reader/main/test/resources/build-config.yaml"
-  );
 });
